# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Furniture API Documentation",
        "description": "API documentation for the Furniture E-commerce Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Current API Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "List user orders"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "shipping_address": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "number"
                                    },
                                    "long": {
                                        "type": "number"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "Cash",
                                            "QR"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/qr/details/{tranId}": {
            "get": {
                "operationId": "2a99a0bebcbd94e772c22decc87cb628",
                "parameters": [
                    {
                        "name": "tranId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get transaction details"
                    }
                }
            }
        },
        "/api/qr/pay/{tranId}": {
            "post": {
                "operationId": "29795cda5fa63bf4771f213396d91945",
                "parameters": [
                    {
                        "name": "tranId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "account_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{invoice_no}/status": {
            "get": {
                "operationId": "4455188e8509f0a051e288232b69a2a6",
                "parameters": [
                    {
                        "name": "invoice_no",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order status"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
    }
}